[CCFPCM-395] mock data, entity unit tests, jest config update
